Auto dotfiles push -- 01_06_2015
Former-commit-id: 2a85322e9640cd61500ee4bd44fb1d3193bdd99c [formerly 6372e6f16467030674769235cefc963b62bd7279]
Former-commit-id: f412d656826e992d272ef5cea7f4de321a3aba08

+++ misc/atom.symlink/packages/linter/lib/linter.coffee
@@ -3,7 +3,10 @@ path = require 'path'
 {Range, Point, BufferedProcess} = require 'atom'
 _ = require 'lodash'
 {XRegExp} = require 'xregexp'
+{CompositeDisposable} = require 'event-kit'
+
 {log, warn} = require './utils'
+
 
 # Public: The base class for linters.
 # Subclasses must at a minimum define the attributes syntax, cmd, and regex.
@@ -50,6 +53,13 @@ class Linter
   constructor: (@editor) ->
     @cwd = path.dirname(@editor.getUri())
 
+    @subscriptions = new CompositeDisposable
+    @subscriptions.add atom.config.observe 'linter.executionTimeout', (x) =>
+      @executionTimeout = x
+
+  destroy: ->
+    @subscriptions.dispose()
+
   # Private: Exists mostly so we can use statSync without slowing down linting.
   # TODO: Do this at constructor time?
   _cachedStatSync: _.memoize (path) ->
@@ -83,6 +93,8 @@ class Linter
     cmd_list = cmd_list.map (cmd_item) ->
       if /@filename/i.test(cmd_item)
         return cmd_item.replace(/@filename/gi, filePath)
+      if /@tempdir/i.test(cmd_item)
+        return cmd_item.replace(/@tempdir/gi, path.dirname(filePath))
       else
         return cmd_item
 
@@ -92,6 +104,9 @@ class Linter
       command: cmd_list[0],
       args: cmd_list.slice(1)
     }
+
+  getReportFilePath: (filePath) ->
+    path.join(path.dirname(filePath), @reportFilePath)
 
   # Private: Provide the node executable path for use when executing a node
   #          linter
@@ -113,6 +128,7 @@ class Linter
 
     dataStdout = []
     dataStderr = []
+    exited = false
 
     stdout = (output) ->
       log 'stdout', output
@@ -123,18 +139,26 @@ class Linter
       dataStderr += output
 
     exit = =>
-      data = if @errorStream is 'stdout' then dataStdout else dataStderr
+      exited = true
+      switch @errorStream
+        when 'file'
+          reportFilePath = @getReportFilePath(filePath)
+          if fs.existsSync reportFilePath
+            data = fs.readFileSync(reportFilePath)
+        when 'stdout' then data = dataStdout
+        else data = dataStderr
       @processMessage data, callback
 
     process = new BufferedProcess({command, args, options,
                                   stdout, stderr, exit})
 
-    # Don't block UI more than 5seconds, it's really annoying on big files
-    timeout_s = 5
-    setTimeout ->
-      process.kill()
-      warn "command `#{command}` timed out after #{timeout_s}s"
-    , timeout_s * 1000
+    # Kill the linter process if it takes too long
+    if @executionTimeout > 0
+      setTimeout =>
+        return if exited
+        process.kill()
+        warn "command `#{command}` timed out after #{@executionTimeout} ms"
+      , @executionTimeout
 
   # Private: process the string result of a linter execution using the regex
   #          as the message builder
@@ -170,7 +194,15 @@ class Linter
     else
       level = @defaultLevel
 
+    # If no line/col is found, assume a full file error
+    # TODO: This conflicts with the docs above that say line is required :(
+    match.line ?= 0
+    match.col ?= 0
+
     return {
+      # TODO: It's confusing that line & col are here since they duplicate info
+      # that's present in the value for range. Consider deprecating line & col
+      # since they're less general than range.
       line: match.line,
       col: match.col,
       level: level,
@@ -187,7 +219,8 @@ class Linter
     match.message
 
   lineLengthForRow: (row) ->
-    return @editor.lineLengthForBufferRow row
+    text = @editor.lineTextForBufferRow row
+    return text?.length or 0
 
   getEditorScopesForPosition: (position) ->
     try
@@ -222,7 +255,6 @@ class Linter
   #   colStart: column to on which to start a higlight (optional)
   #   colEnd: column to end highlight (optional)
   computeRange: (match) ->
-    match.line ?= 0 # Assume if no line is found that it denotes a full file error.
 
     decrementParse = (x) ->
       Math.max 0, parseInt(x) - 1
@@ -236,7 +268,6 @@ class Linter
       log "ignoring #{match} - it's longer than the buffer"
       return null
 
-    match.col ?=  0
     unless match.colStart
       position = new Point(rowStart, match.col)
       scopes = @getEditorScopesForPosition(position)
